Protector Test and Protector Fix

+++ Assets/Scripts/App/Managers/PvPManager.cs
@@ -338,7 +338,7 @@ namespace Loom.ZombieBattleground
                         }
                         if (matchCanStart)
                         {
-                            MersenneTwister.RandomInit(playerActionEvent.Match.RandomSeed);
+                            MersenneTwister.RandomInit((uint)playerActionEvent.Match.RandomSeed);
                             MatchingStartedActionReceived?.Invoke();
                         }
                         break;
